requirements.txt
pycryptodomex
hashlib
requests
python-dotenv